requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.2.0
tqdm>=4.65.0
tensorboard>=2.13.0
pyyaml>=6.0

timm>=0.9.0 
einops>=0.6.0  

ultralytics>=8.0.0  
lap>=0.4.0  

pandas>=2.0.0
h5py>=3.8.0

easydict>=1.10
termcolor>=2.3.0
tabulate>=0.9.0

pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
ipython>=8.12.0
jupyter>=1.0.0
